+++ src/main/java/com/nastya/model/ExchangeRate.java
@@ -1,15 +1,17 @@
 package com.nastya.model;
 
+import java.math.BigDecimal;
+
 public class ExchangeRate {
     private int id;
     private Currency baseCurrency;
     private Currency targetCurrency;
-    private double rate;
+    private BigDecimal rate;
 
     public ExchangeRate() {
     }
 
-    public ExchangeRate(int id, Currency baseCurrency, Currency targetCurrency, double rate) {
+    public ExchangeRate(int id, Currency baseCurrency, Currency targetCurrency, BigDecimal rate) {
         this.id = id;
         this.baseCurrency = baseCurrency;
         this.targetCurrency = targetCurrency;
@@ -28,7 +30,7 @@ public class ExchangeRate {
         return targetCurrency;
     }
 
-    public double getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
 
@@ -44,7 +46,7 @@ public class ExchangeRate {
         this.targetCurrency = targetCurrency;
     }
 
-    public void setRate(double rate) {
+    public void setRate(BigDecimal rate) {
         this.rate = rate;
     }
 }
